More deleting and renaming
